- Remove debug-only Q_QUERY_EXEC_TIME status var.
- Update code for 5.7.8 idioms, notably ULONGLONGMAX replaced with
  ULLONGMAX, extra scope field for plugin status vars, removal of
  DYNAMIC_ARRAY in C++ code, removal of UNIV_SYNC_DEBUG, ib_logf
  removal.

+++ sql/rpl_info_table_access.cc
@@ -1,4 +1,4 @@
-/* Copyright (c) 2010, 2014, Oracle and/or its affiliates. All rights reserved.
+/* Copyright (c) 2010, 2015, Oracle and/or its affiliates. All rights reserved.
 
    This program is free software; you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
@@ -14,9 +14,12 @@
    Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301  USA */
 
 #include "rpl_info_table_access.h"
-#include "rpl_utility.h"
-#include "handler.h"
-#include "sql_parse.h"
+
+#include "rpl_info_values.h" // Rpl_info_values
+#include "sql_base.h"       // MYSQL_OPEN_IGNORE_FLUSH
+#include "sql_class.h"      // THD
+#include "sql_parse.h"      // mysql_reset_thd_for_next_command
+#include "table.h"          // TABLE
 
 
 void Rpl_info_table_access::before_open(THD *thd)
@@ -105,7 +108,7 @@ enum enum_return_id Rpl_info_table_access::find_info(Rpl_info_values *field_valu
     DBUG_RETURN(ERROR_ID);
   }
 
-  keyinfo= table->s->key_info + (uint) table->s->primary_key;
+  keyinfo= table->s->key_info + table->s->primary_key;
   for (uint idx= 0; idx < keyinfo->user_defined_key_parts; idx++)
   {
     uint fieldnr= keyinfo->key_part[idx].fieldnr - 1;
