+++ sql/uniques.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2001, 2014, Oracle and/or its affiliates. All rights reserved.
+/* Copyright (c) 2001, 2015, Oracle and/or its affiliates. All rights reserved.
 
    This program is free software; you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
@@ -16,10 +16,9 @@
 #ifndef UNIQUES_INCLUDED
 #define UNIQUES_INCLUDED
 
-#include "sql_alloc.h"
-#include "prealloced_array.h"
-#include "sql_array.h"
-#include "sql_sort.h"
+#include "my_tree.h"            // TREE
+#include "prealloced_array.h"   // Prealloced_array
+#include "sql_alloc.h"          // Sql_alloc
 
 class Cost_model_table;
 
@@ -72,7 +71,7 @@ public:
   {
     ulonglong max_elems_in_tree=
       (max_in_memory_size / ALIGN_SIZE(sizeof(TREE_ELEMENT)+key_size));
-    return 1 + nkeys/max_elems_in_tree;
+    return 1 + static_cast<size_t>(nkeys/max_elems_in_tree);
   }
 
   void reset();
